refactor: remove dead createGlobalVariableValidation + extract getNextName
- Remove unused createGlobalVariableValidation from validation/variables.ts
  (replaced by createVariableValidation for both local and global vars)
- Extract duplicated getNextName helper from task-editor and
  instances-editor into shared src/frontend/utils/next-name.ts
